infrastructure/etl/odoo-sync/requirements.txt
# Odoo → Supabase ETL Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
supabase>=2.0.0
pydantic>=2.0.0
